Setup basic structure
The basic structure is in place. Everything compiles, but the main
program uses sysfs to access the sht45 device.

There is a launch,json files and some tasks that allow running the
debugger. THat all seems to work. The tasks run some scripts that
are specific to my environment and would need to be changed at any
other site.