Got user space i2c commands working
This is a check-in to save that I got the i2c commands to work.
There is a lot of things that need to get corrected. The data
agrees with the data accessed from /sysfs within 1 % so seems
correct.

1. Had to add some sleeps() around the ioctl's or the debugger would
get very confused. These sleeps need be shorter

2. Need to implement the CRC check in the retreived data.

3. Need better error handling.

Note: I tried to use expected, but it complained about no template or
something.

Still a lot of work here but we now know the pieces work.