Fixed up the elapsed time check
We don't need to hit the i2c bus on every request. The data just
doesn't change that fast. Right now we only get new data if it
has been more than a second since we last got a a measurement.
We could probably be OK with every minute or even every 5 minutes.
For now we stick with a second.