Add error checking
Added the -std-c++23 option in order to be able to use expected class.
THe expected class is used when a routine returns a value but has
the possibilty of having an error.

Still need to standardize the error codes. Right now an error is just
an integer that increments within a routine. AN error value of 0 means
success.