Add an I2cBus class to handle I2C transfers
The devices make calls to the I2cBus routines to actually transfer
data. There are two types of exchnages.

The command based is used by the sht4x. You send a one byte command
and then the next read to the slave address will return the result.

The register treats the sub-addresses as indivudual registers. You
can write or read to any register that allows it. These also
support multi-register reads and writes. For this to work the
device has to auto-increment after every read or write.