Use milli-celsius as a base value and fix chrono
I decided to store temperature in millicelsius which agress with
the UNIX system implementation.

I fixed up the use of chrono

I added a fix to workaround the fact that the device always returns
a temperature value of zero degrees Celsius on first read after
a power cycle of the device.