Use locks and fix overrun management
We use guard_lock any place where we want to make a sequence of
data transfers. This prevents another thread from interrupting the
sequence of commands.

We guard a recursive_lock. This allows one routine that needs to
make transfers to call another routine that needs to make transfers.
For example reset is a stand alone routine that can be called directly.
We may also need to reset the chip in the getMeasurement if an
error happens and we may want to reset the chip and try again.

The chip sometimes sets the temperature overrun flag. This was being
treated as an error and no data read was made. By not reading the
data value the overrun condition was never cleared so the condition
never cleared.

The overrun means that a new data copied over an older one. Since
we are running in one-shot mode we don't really care. Getting the
most recent value is actually what we want.