Create global devices list
The devices list keeps the device specific data that can be used by
multiple active instances. The device data is put on a list for
all the different devices that have ever been opened by an instance.

The constructor makes sanity checks of the data passed. If the device
data already exists on the devices list it just attaches that device
data to the instance.

If the device is not found on the list the constructor creates a new
device data structure and tries to do a whoami. If it is successful
then it adds the the device data to the devices list.

If you can't do a whoami then the device_data memory is freed and
the device_data_ is set to nullptr. All other routines that will
access the device check if device_data_ is nullptr. If it is they
all return ENODEV.