Add class to report to Weather Underground
Added the WeatherUnderground class so we could report the data
to Weather Underground. I had to get an account for Weather Underground
and create a PWS, Personal Weather Station. They provide and ID number
and password. I was able to successfully send data to the Weather
Underground and see the data appear on their PWS for my account.