Add a script to get manual packages from apt list
Added load_manual_packages which takes a file that is the output
of "apt list --installed".

To duplicate a package list on the original host run

$ apt list --installed > pkgs_installed.lst

Then copy this machine to tanother host that has just had the OS
loaded on it. You can then run this scrpit to load all the manually
loaded packages on the other machine.