Switch over to using the new units paradigm
This current version only prints the value and doesn't actually
send to Wetaher Underground and has a qucik cyce time.

Switched to a paradigm of having each unit type be it's own class.
We convert from one unit to another through implicit casting.

Esssentially, float values are converted to a uint64_t base value. For
temperature the base value represents a millicelsius degrees. For
pressure the base value is 1 thousandth of a millbar.

The value() routine of each type converts the base value to the
appropriate value for the unit type.

All comparison operators are based off of comparing the base values.

There are also conversion operators that provide impplicit casting from
one type of unit to another.