Change the systemd access model
Use a model that uses objects indicated in the man page at:

https://www.freedesktop.org/software/systemd/man/latest/org.freedesktop.systemd1.html

It breaks things up into objects. Each object has it's own set of
methods and properties. So, it's a way to be able to get any methods
and properties by using the correct object type.

See the README.md for more details.