Add support for Ads1015 12 bit ADC on I2c bus.
We add support for the ads1015 ADC 12 bit controller. This has up to
4 differential or 4 single-ended ADC inputs or some combination
thereof.

At this point I have only been able to test reading the configuration
register. That seems to work. I need to stage having a wind speed
sensor connected to the device in order to test reading actual data.