Add the anomometer device that uses ads1015 ADC
We add the anomometer weather device. It uses the Ads1015 ADC device.

I had to make some changes to ads1015, but it seems I am getting
resonable values when using a power supply to test the voltage
value.

I have not connected to actual anomometer to see how that works.