Use inheritance in speed instead of variants.
By creating a speed class that is used as a base for all the
speed unit classes it makes the code simpler. I can put all the
operator in speed class. I also put all the implicit cast operators
in speed. THis makes all the speed unit classes much simpler.