Fix up units casting, update to report wind info
I couldn't have a constructor for the units for both float and int64_t.
It generated ambiguous constructor errors. So, had to add a setBase()
method that I made private. Since it was private I had to add the base
class as a friend to the derived class.

THere may be a differentway to create a derived class from a base class
for implicit casting that doesn't require the setBase() method. I will
have to check it out.