Fix wind history sorting issue and change looping.
I was sorting the history in the wrong order. THe 2 minute average
was incorrect as a result. Once the order was changed the problem
went away.

Also, wind speed requires much more frequent looping than just reporting
data. So, we have a datagathering loop where we gather wind info and
update the wind history. We only need to gather the other data
when we need to do a report.