Create just one configuration class.
There are two configuration files. One can be modified the other is
read only. The read only one is fixed and only chnages via an upgrade.
The configurable one has values that can be changed while the
process is running. Variables that make sense to be chnaged are in the
configurable file. One class WeatherStationConfig gives the program
access to all variables across both files.