Separate packages needed for dev or running
Make one list for the packages needed for development and a separate
list of packages needed for running. In the future we may be able
to only load the runtime packages if the host won't be used to
make an SDK for building.

Also added modbus packages in expectation of the EcoWitt LN90LP device.