Add google testing support
I had to add gtest libraries and headers to the project. This involved
building the googletests and copying them to /usr/include and /usr/lib.
I then moved these manually to the development host SDK.

I had to add enable_testing() to the CMakeLists.txt file to enable the
testing framework. I got erros about no DartConfig.tcl file found.
I read that Dart was deprecated and to use include(Dart) to add it to
CMake. I added this line but it did not fix the error. I am not sure
if this is needed or not.

I may have needed to add it once to get the DartConfig.tcl file
generated, but I am not sure. I will have to look into this more.