Add ecowitt device support and other sundry things
Added the ecowitt device support.

Create weather devices like thermometer, barometer, hygrometer etc..
These are created by using the std::function that allows me to define
a set a functions from one of the device drivers to these weather
devices. For example I can create a thermometer that uses the ecowitt
device that gets the temperature and then another thermometer that
uses one of the devices on the I2C bus to get the temperature.

I created Direction units in order to get the wind direction from
ecowitt device.

Since I need a history for both the wind speed and the wind direction
the history was generalized with the use of templates.

Started adding some unit tests for the units. Ran into a problem though
where I can't seem to have a generic launch.json and it's PreLaunch task
that allows me to run and debug any test. I keep having to change
them depending on which test I want to run.